Remove the unit default collection

+++ db/migrate/20230215191407_remove_bitstreams_submitted_for_ingest.rb
@@ -1,5 +1,5 @@
 class RemoveBitstreamsSubmittedForIngest < ActiveRecord::Migration[7.0]
   def change
-    remove_column :bitstreams, :submitted_for_ingest
+    remove_column :bitstreams, :submitted_for_ingest if column_exists? :bitstreams, :submitted_for_ingest
   end
 end
